backend/pyproject.toml
[project]
name = "rbsoftware-backend"
version = "0.1.0"
description = "RBSoftware backend foundation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "sqlmodel>=0.0.16",
  "alembic>=1.13.1",
  "pydantic-settings>=2.2.1",
  "pymysql>=1.1.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=3.2.0,<4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.4.2",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --disable-warnings"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
